refactor: reorganize ride tracking components by user role
