+++ frontend/src/app/features/ride-tracking/components/driver/stop-ride/stop-ride.component.scss
