+++ frontend/src/app/features/ride-tracking/components/shared/panic/panic.component.ts
